New Fix for Corrupted SD Card
There are several places in the code where it sets the internal SD clock to operate at 50 MHz (if SD card is capable).  Unfortunately, hardware/firmware on the camera doesn't reliably work above 25 MHz.  The result of which is that *sometimes* writes to the SD card operating at 50 MHz will corrupt it.  In my first fix, I addresed one of these entry points.  In this new fix, I think I cover them all.