Major Release 20250321R
A major release, including new functions, and generalization of  previous functions.

I have also updated the firmare for the HP5 cameras so that they are based on the more recent factory firmware, version "M06170F." I'm not sure what's new in this factory release, although I don't believe it addresses the "corrupt SD card" bug.  My fix for this bug remains in place.

I added support for the BTC-7A (Recon Force Advantage) model.  This firmware is based on the more recent model for the BTC-7E, and thus inherits the "capture timer" feature Bronwning introduced with the "Elite" models.

I added Several new features, described in more detail below.  These include:

* Saving and Restoring Custom Settings

* Native Photo Quality

* External Trigger

I have also removed the "Extended SD Card" menu.  This was a lightly used, and likely obsolete function, and I needed to retire it to make room for the custom configuration code, above.

As of 2025-03-18, there is a known issue that prevents certain newer HP5 cameras from downloading this firmware (or Browning firmware, for that matter).  If the "YES" option isn't available in the "FW UPGRADE" menu, you may have one of these cameras.